Extracts common code
